+++ src/main/java/com/digitechbb/smsv1/services/ProfessorServiceImpl.java
@@ -7,16 +7,22 @@ import com.digitechbb.smsv1.repositories.ProfessorRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+
 @Service
 @AllArgsConstructor
-public class ProfessorServiceImpl implements ProfessorService{
+public class ProfessorServiceImpl implements ProfessorService {
     private final ProfessorRepository professorRepository;
     private final ProfessorMapper professorMapper;
 
     @Override
     public ProfessorDto save(ProfessorDto professorDto) {
+//        try{
+//
+//        }catch (Exception e ) {
+//            e.printStackTrace();
+//        }
         return professorMapper.toDto(professorRepository.save(professorMapper.toProfessor(professorDto)));
     }
 
@@ -35,25 +41,25 @@ public class ProfessorServiceImpl implements ProfessorService{
     @Override
     public ProfessorDto getProfessor(Long id) {
         Professor professor = professorRepository.findById(id).orElse(null);
-        if (professor==null)
+        if (professor == null)
             return null;
         return professorMapper.toDto(professor);
     }
 
     @Override
     public List<ProfessorDto> getAllProfessors() {
-        List<Professor> professorList = professorRepository.findAll();
-        if (professorList==null)
-            return null;
-        List<ProfessorDto> professorDtoList = new ArrayList<>();
-        for (Professor p: professorList) {
-            professorDtoList.add(professorMapper.toDto(p));
-        }
-        return professorDtoList;
+        return professorRepository.findAll().stream()
+                .map(professor -> professorMapper.toDto(professor)).collect(Collectors.toList());
     }
 
     @Override
-    public void deleteProfessorById(Long id) {
-        professorRepository.deleteById(id);
+    public boolean deleteProfessorById(Long id) {
+        Professor professor = professorRepository.findById(id).orElse(null);
+        if (professor != null) {
+            professorRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
